finalizado modulo 4

+++ projetos/automated-tests-with-java/src/main/java/br/com/daniel/SimpleMath.java
@@ -25,7 +25,7 @@ public class SimpleMath {
 	}
 
 	public Double squareRoot(Double number) {
-		return (Double) Math.sqrt(number);
+		return Math.sqrt(number);
 	}
 
 }
